Ingreso Comunidad Detalle funcionando

+++ KermesseApp/Models/tbl_bonos.cs
@@ -22,7 +22,7 @@ namespace KermesseApp.Models
     
         public int id_bono { get; set; }
         public string nombre { get; set; }
-        public double valor { get; set; }
+        public decimal valor { get; set; }
         public int estado { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
